@@ -1,3 +1,5 @@
-export * from './analytics';
-export * from './analytics.module';
-export * from './analytics.service';
+export { Analytics, AnalyticsInstances, analyticInstance$ } from './analytics';
+export { provideAnalytics, AnalyticsModule } from './analytics.module';
+export * from './firebase';
+export * from './screen-tracking.service';
+export * from './user-tracking.service';
