@@ -1,63 +1,30 @@
-import { Inject, Injectable, InjectionToken, NgZone, Optional } from '@angular/core';
-import { from, Observable, of } from 'rxjs';
-import { map, observeOn, shareReplay, switchMap, tap } from 'rxjs/operators';
-import {
-  FIREBASE_APP_NAME,
-  FIREBASE_OPTIONS,
-  FirebaseAppConfig,
-  FirebaseOptions,
-  ɵAngularFireSchedulers,
-  ɵfirebaseAppFactory,
-  ɵlazySDKProxy,
-  ɵPromiseProxy
-} from '@angular/fire';
-import { functions } from 'firebase/app';
+import { Functions as FirebaseFunctions } from 'firebase/functions';
+import { ɵgetAllInstancesOf } from '@angular/fire';
+import { from, timer } from 'rxjs';
+import { concatMap, distinct } from 'rxjs/operators';
 
-export const ORIGIN = new InjectionToken<string>('angularfire2.functions.origin');
-export const REGION = new InjectionToken<string>('angularfire2.functions.region');
+// see notes in core/firebase.app.module.ts for why we're building the class like this
+// tslint:disable-next-line:no-empty-interface
+export interface Functions extends FirebaseFunctions {}
 
-// override httpsCallable for compatibility with 5.x
-export interface AngularFireFunctions extends Omit<ɵPromiseProxy<functions.Functions>, 'httpsCallable'> {
-}
-
-@Injectable({
-  providedIn: 'any'
-})
-export class AngularFireFunctions {
-
-  public readonly httpsCallable: <T = any, R = any>(name: string) => (data: T) => Observable<R>;
-
-  constructor(
-    @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
-    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig: string | FirebaseAppConfig | null | undefined,
-    zone: NgZone,
-    @Optional() @Inject(REGION) region: string | null,
-    @Optional() @Inject(ORIGIN) origin: string | null
-  ) {
-    const schedulers = new ɵAngularFireSchedulers(zone);
-
-    const functions = of(undefined).pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(() => import('firebase/functions')),
-      map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
-      map(app => app.functions(region || undefined)),
-      tap(functions => {
-        if (origin) {
-          functions.useFunctionsEmulator(origin);
-        }
-      }),
-      shareReplay({ bufferSize: 1, refCount: false })
-    );
-
-    this.httpsCallable = <T = any, R = any>(name: string) =>
-      (data: T) => from(functions).pipe(
-        observeOn(schedulers.insideAngular),
-        switchMap(functions => functions.httpsCallable(name)(data)),
-        map(r => r.data as R)
-      );
-
-    return ɵlazySDKProxy(this, functions, zone);
-
+export class Functions {
+  constructor(functions: FirebaseFunctions) {
+    return functions;
   }
-
 }
+
+export const FUNCTIONS_PROVIDER_NAME = 'functions';
+
+// tslint:disable-next-line:no-empty-interface
+export interface FunctionsInstances extends Array<FirebaseFunctions> {}
+
+export class FunctionsInstances {
+  constructor() {
+    return ɵgetAllInstancesOf<FirebaseFunctions>(FUNCTIONS_PROVIDER_NAME);
+  }
+}
+
+export const functionInstance$ = timer(0, 300).pipe(
+  concatMap(() => from(ɵgetAllInstancesOf<FirebaseFunctions>(FUNCTIONS_PROVIDER_NAME))),
+  distinct(),
+);
