@@ -1,2 +1,4 @@
-export * from './functions';
-export * from './functions.module';
+export { provideFunctions, FunctionsModule } from './functions.module';
+export { Functions, FunctionsInstances, functionInstance$ } from './functions';
+export * from './rxfire';
+export * from './firebase';
