@@ -1,73 +1,43 @@
 import { TestBed } from '@angular/core/testing';
-import { AngularFireModule, FIREBASE_APP_NAME, FIREBASE_OPTIONS, FirebaseApp } from '@angular/fire';
-import { AngularFireMessaging, AngularFireMessagingModule } from './public_api';
+import { FirebaseApp, provideFirebaseApp, getApp, initializeApp, deleteApp } from '@angular/fire/app';
+import { Messaging, provideMessaging, getMessaging, isSupported } from '@angular/fire/messaging';
 import { COMMON_CONFIG } from '../test-config';
-import { rando } from '../firestore/utils.spec';
+import { rando } from '../utils';
 
-describe('AngularFireMessaging', () => {
+describe('Messaging', () => {
   let app: FirebaseApp;
-  let afm: AngularFireMessaging;
+  let messaging: Messaging;
+  let providedMessaging: Messaging;
+  let appName: string;
 
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
-        AngularFireMessagingModule
-      ]
+  beforeAll(done => {
+    // The APP_INITIALIZER that is making isSupported() sync for DI may not
+    // be done evaulating by the time we inject from the TestBed. We can
+    // ensure correct behavior by waiting for the (global) isSuppported() promise
+    // to resolve.
+    isSupported().then(() => done());
+  });
+
+  describe('single injection', () => {
+
+    beforeEach(() => {
+        appName = rando();
+        TestBed.configureTestingModule({
+            imports: [
+                provideFirebaseApp(() => initializeApp(COMMON_CONFIG, appName)),
+                provideMessaging(() => {
+                    providedMessaging = getMessaging(getApp(appName));
+                    return providedMessaging;
+                }),
+            ],
+        });
+        app = TestBed.inject(FirebaseApp);
+        messaging = TestBed.inject(Messaging);
     });
 
-    app = TestBed.inject(FirebaseApp);
-    afm = TestBed.inject(AngularFireMessaging);
-  });
-
-  afterEach(() => {
-    app.delete();
-  });
-
-  it('should be exist', () => {
-    expect(afm instanceof AngularFireMessaging).toBe(true);
-  });
-
-  it('should have the FCM instance', () => {
-    expect(afm.deleteToken).toBeDefined();
-  });
-
-});
-
-const FIREBASE_APP_NAME_TOO = (Math.random() + 1).toString(36).substring(7);
-
-describe('AngularFireMessaging with different app', () => {
-  let app: FirebaseApp;
-  let afm: AngularFireMessaging;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
-        AngularFireMessagingModule
-      ],
-      providers: [
-        { provide: FIREBASE_APP_NAME, useValue: FIREBASE_APP_NAME_TOO },
-        { provide: FIREBASE_OPTIONS, useValue: COMMON_CONFIG }
-      ]
-    });
-
-    app = TestBed.inject(FirebaseApp);
-    afm = TestBed.inject(AngularFireMessaging);
-  });
-
-  afterEach(() => {
-    app.delete();
-  });
-
-  describe('<constructor>', () => {
-
-    it('should be an AngularFireMessaging type', () => {
-      expect(afm instanceof AngularFireMessaging).toEqual(true);
-    });
-
-    it('should have the FCM instance', () => {
-      expect(afm.deleteToken).toBeDefined();
+    it('should be injectable', () => {
+        expect(providedMessaging).toBeTruthy();
+        expect(messaging).toEqual(providedMessaging);
     });
 
   });
