@@ -1,75 +1,44 @@
 import { TestBed } from '@angular/core/testing';
-import { AngularFireModule, FIREBASE_APP_NAME, FIREBASE_OPTIONS, FirebaseApp } from '@angular/fire';
-import { AngularFireRemoteConfig, AngularFireRemoteConfigModule, DEFAULTS, SETTINGS } from './public_api';
+import { FirebaseApp, provideFirebaseApp, getApp, initializeApp, deleteApp } from '@angular/fire/app';
+import { RemoteConfig, provideRemoteConfig, getRemoteConfig, isSupported } from '@angular/fire/remote-config';
 import { COMMON_CONFIG } from '../test-config';
-import { rando } from '../firestore/utils.spec';
+import { rando } from '../utils';
 
-describe('AngularFireRemoteConfig', () => {
+describe('RemoteConfig', () => {
   let app: FirebaseApp;
-  let rc: AngularFireRemoteConfig;
+  let remoteConfig: RemoteConfig;
+  let providedRemoteConfig: RemoteConfig;
+  let appName: string;
 
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
-        AngularFireRemoteConfigModule
-      ]
+  beforeAll(done => {
+    // The APP_INITIALIZER that is making isSupported() sync for DI may not
+    // be done evaulating by the time we inject from the TestBed. We can
+    // ensure correct behavior by waiting for the (global) isSuppported() promise
+    // to resolve.
+    isSupported().then(() => done());
+  });
+
+  describe('single injection', () => {
+
+    beforeEach(() => {
+        appName = rando();
+        TestBed.configureTestingModule({
+            imports: [
+                provideFirebaseApp(() => initializeApp(COMMON_CONFIG, appName)),
+                provideRemoteConfig(() => {
+                    providedRemoteConfig = getRemoteConfig(getApp(appName));
+                    return providedRemoteConfig;
+                }),
+            ],
+        });
+        app = TestBed.inject(FirebaseApp);
+        remoteConfig = TestBed.inject(RemoteConfig);
     });
 
-    app = TestBed.inject(FirebaseApp);
-    rc = TestBed.inject(AngularFireRemoteConfig);
-  });
-
-  afterEach(() => {
-    app.delete();
-  });
-
-  it('should be exist', () => {
-    expect(rc instanceof AngularFireRemoteConfig).toBe(true);
-  });
-
-  it('should have the Firebase Functions instance', () => {
-    expect(rc.getValue).toBeDefined();
-  });
-
-});
-
-const FIREBASE_APP_NAME_TOO = (Math.random() + 1).toString(36).substring(7);
-
-describe('AngularFireRemoteConfig with different app', () => {
-  let app: FirebaseApp;
-  let rc: AngularFireRemoteConfig;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        AngularFireModule.initializeApp(COMMON_CONFIG, rando()),
-        AngularFireRemoteConfigModule
-      ],
-      providers: [
-        { provide: FIREBASE_APP_NAME, useValue: FIREBASE_APP_NAME_TOO },
-        { provide: FIREBASE_OPTIONS, useValue: COMMON_CONFIG },
-        { provide: SETTINGS, useValue: {} },
-        { provide: DEFAULTS, useValue: {} }
-      ]
-    });
-
-    app = TestBed.inject(FirebaseApp);
-    rc = TestBed.inject(AngularFireRemoteConfig);
-  });
-
-  afterEach(() => {
-    app.delete();
-  });
-
-  describe('<constructor>', () => {
-
-    it('should be an AngularFireAuth type', () => {
-      expect(rc instanceof AngularFireRemoteConfig).toEqual(true);
-    });
-
-    it('should have the Firebase Functions instance', () => {
-      expect(rc.getValue).toBeDefined();
+    it('should be injectable', () => {
+        expect(providedRemoteConfig).toBeTruthy();
+        expect(remoteConfig).toEqual(providedRemoteConfig);
+        expect(remoteConfig.app).toEqual(app);
     });
 
   });
