@@ -1,7 +1,8 @@
-import { experimental, JsonObject, logging } from '@angular-devkit/core';
+import { JsonObject, logging } from '@angular-devkit/core';
 import { BuilderContext, BuilderRun, ScheduleOptions, Target } from '@angular-devkit/architect';
 import { BuildTarget, FirebaseDeployConfig, FirebaseTools, FSHost } from '../interfaces';
-import deploy, { deployToFunction } from './actions';
+import deploy, { deployToFunction } from '@angular/fire/schematics/deploy/actions';
+import { join } from 'path';
 import 'jasmine';
 
 let context: BuilderContext;
@@ -10,9 +11,22 @@ let fsHost: FSHost;
 
 const FIREBASE_PROJECT = 'ikachu-aa3ef';
 const PROJECT = 'pirojok-project';
-const BUILD_TARGET: BuildTarget = {
+const STATIC_BUILD_TARGET: BuildTarget = {
   name: `${PROJECT}:build:production`
 };
+
+const FIREBASE_TOKEN = 'kkasllkascnkjnskjsdcskdckskdksdkjc';
+
+const SERVER_BUILD_TARGET: BuildTarget = {
+  name: `${PROJECT}:server:production`
+};
+
+const login = () => Promise.resolve({ user: { email: 'foo@bar.baz' }});
+login.list = () => Promise.resolve([{ user: { email: 'foo@bar.baz' }}]);
+login.add = () => Promise.resolve([{ user: { email: 'foo@bar.baz' }}]);
+login.use = () => Promise.resolve('foo@bar.baz');
+
+const workspaceRoot = join('home', 'user');
 
 const initMocks = () => {
   fsHost = {
@@ -21,20 +35,42 @@ const initMocks = () => {
     renameSync(_: string, __: string) {
     },
     writeFileSync(_: string, __: string) {
-    }
+    },
+    copySync(_: string, __: string) {
+    },
+    removeSync(_: string) {
+    },
+    existsSync(_: string) {
+      return false;
+    },
   };
 
   firebaseMock = {
-    login: () => Promise.resolve(),
+    login,
     projects: {
-      list: () => Promise.resolve([])
+      list: () => Promise.resolve([]),
+      create: () => Promise.reject(),
+    },
+    apps: {
+      list: () => Promise.resolve([]),
+      create: () => Promise.reject(),
+      sdkconfig: () => Promise.resolve({ fileName: '_', fileContents: '', sdkConfig: {}, }),
+    },
+    hosting: {
+      sites: {
+        list: () => Promise.resolve({sites: []}),
+        create: () => Promise.reject(),
+      }
     },
     deploy: (_: FirebaseDeployConfig) => Promise.resolve(),
     use: () => Promise.resolve(),
     logger: {
-      add: () => {
+      add: () => { },
+      logger: {
+        add: () => { }
       }
     },
+    cli: { version: () => '9.0.0' },
     serve: () => Promise.resolve()
   };
 
@@ -53,7 +89,13 @@ const initMocks = () => {
     id: 1,
     logger: new logging.NullLogger() as any,
     workspaceRoot: 'cwd',
-    getTargetOptions: (_: Target) => Promise.resolve({}),
+    getTargetOptions: async (target: Target) => {
+      if (target.target === 'build') {
+        return { outputPath: 'dist/browser' };
+      } else if (target.target === 'server') {
+        return { outputPath: 'dist/server' };
+      }
+    },
     reportProgress: (_: number, __?: number, ___?: string) => {
     },
     reportStatus: (_: string) => {
@@ -65,32 +107,27 @@ const initMocks = () => {
   } as any);
 };
 
-
-const projectTargets: experimental.workspace.WorkspaceTool = {
-  build: {
-    options: {
-      outputPath: 'dist/browser'
-    }
-  },
-  server: {
-    options: {
-      outputPath: 'dist/server'
-    }
-  }
-};
-
 describe('Deploy Angular apps', () => {
   beforeEach(() => initMocks());
 
   it('should call login', async () => {
-    const spy = spyOn(firebaseMock, 'login');
-    await deploy(firebaseMock, context, projectTargets, [BUILD_TARGET], FIREBASE_PROJECT, false, false);
+    const spy = spyOn(firebaseMock, 'login').and.resolveTo({ email: 'foo@bar.baz' });
+    await deploy(
+      firebaseMock, context, STATIC_BUILD_TARGET, undefined,
+      undefined, undefined, { projectId: FIREBASE_PROJECT, preview: false }
+    );
     expect(spy).toHaveBeenCalled();
+  });
+
+  it('should not call login', async () => {
+    const spy = spyOn(firebaseMock, 'login');
+    await deploy(firebaseMock, context, STATIC_BUILD_TARGET, undefined,  undefined, undefined, { preview: false }, FIREBASE_TOKEN);
+    expect(spy).not.toHaveBeenCalled();
   });
 
   it('should invoke the builder', async () => {
     const spy = spyOn(context, 'scheduleTarget').and.callThrough();
-    await deploy(firebaseMock, context, projectTargets, [BUILD_TARGET], FIREBASE_PROJECT, false, false);
+    await deploy(firebaseMock, context, STATIC_BUILD_TARGET, undefined,  undefined, undefined, { preview: false });
     expect(spy).toHaveBeenCalled();
     expect(spy).toHaveBeenCalledWith({
       target: 'build',
@@ -105,27 +142,29 @@ describe('Deploy Angular apps', () => {
       options: {}
     };
     const spy = spyOn(context, 'scheduleTarget').and.callThrough();
-    await deploy(firebaseMock, context, projectTargets, [buildTarget], FIREBASE_PROJECT, false, false);
+    await deploy(firebaseMock, context, buildTarget, undefined,  undefined, undefined, { preview: false });
     expect(spy).toHaveBeenCalled();
     expect(spy).toHaveBeenCalledWith({ target: 'prerender', project: PROJECT }, {});
   });
 
   it('should invoke firebase.deploy', async () => {
     const spy = spyOn(firebaseMock, 'deploy').and.callThrough();
-    await deploy(firebaseMock, context, projectTargets, [BUILD_TARGET], FIREBASE_PROJECT, false, false);
+    await deploy(firebaseMock, context, STATIC_BUILD_TARGET, undefined,  undefined, undefined, { preview: false }, FIREBASE_TOKEN);
     expect(spy).toHaveBeenCalled();
     expect(spy).toHaveBeenCalledWith({
       cwd: 'cwd',
-      only: 'hosting:' + PROJECT
+      only: 'hosting:' + PROJECT,
+      token: FIREBASE_TOKEN,
+      nonInteractive: true,
+      projectRoot: 'cwd',
     });
   });
 
   describe('error handling', () => {
     it('throws if there is no firebase project', async () => {
       try {
-        await deploy(firebaseMock, context, projectTargets, [BUILD_TARGET], undefined, false, false);
+        await deploy(firebaseMock, context, STATIC_BUILD_TARGET, undefined, undefined, undefined, { preview: false  });
       } catch (e) {
-        console.log(e);
         expect(e.message).toMatch(/Cannot find firebase project/);
       }
     });
@@ -133,7 +172,7 @@ describe('Deploy Angular apps', () => {
     it('throws if there is no target project', async () => {
       context.target = undefined;
       try {
-        await deploy(firebaseMock, context, projectTargets, [BUILD_TARGET], FIREBASE_PROJECT, false, false);
+        await deploy(firebaseMock, context, STATIC_BUILD_TARGET, undefined, undefined, undefined, { preview: false });
       } catch (e) {
         expect(e.message).toMatch(/Cannot execute the build target/);
       }
@@ -146,34 +185,106 @@ describe('universal deployment', () => {
 
   it('should create a firebase function', async () => {
     const spy = spyOn(fsHost, 'writeFileSync');
-    await deployToFunction(firebaseMock, context, '/home/user', projectTargets, false, fsHost);
+    await deployToFunction(
+      firebaseMock,
+      context,
+      workspaceRoot,
+      STATIC_BUILD_TARGET,
+      SERVER_BUILD_TARGET,
+      { preview: false  },
+      undefined,
+      fsHost
+    );
 
     expect(spy).toHaveBeenCalledTimes(2);
 
     const packageArgs = spy.calls.argsFor(0);
     const functionArgs = spy.calls.argsFor(1);
 
-    expect(packageArgs[0]).toBe('dist/package.json');
-    expect(functionArgs[0]).toBe('dist/index.js');
+    expect(packageArgs[0]).toBe(join(workspaceRoot, 'dist', 'package.json'));
+    expect(functionArgs[0]).toBe(join(workspaceRoot, 'dist', 'index.js'));
+  });
+
+  it('should create a firebase function (new)', async () => {
+    const spy = spyOn(fsHost, 'writeFileSync');
+    await deployToFunction(
+      firebaseMock,
+      context,
+      workspaceRoot,
+      STATIC_BUILD_TARGET,
+      SERVER_BUILD_TARGET,
+      { preview: false, outputPath: join('dist', 'functions') },
+      undefined,
+      fsHost
+    );
+
+    expect(spy).toHaveBeenCalledTimes(2);
+
+    const packageArgs = spy.calls.argsFor(0);
+    const functionArgs = spy.calls.argsFor(1);
+
+    expect(packageArgs[0]).toBe(join(workspaceRoot, 'dist', 'functions', 'package.json'));
+    expect(functionArgs[0]).toBe(join(workspaceRoot, 'dist', 'functions', 'index.js'));
   });
 
   it('should rename the index.html file in the nested dist', async () => {
     const spy = spyOn(fsHost, 'renameSync');
-    await deployToFunction(firebaseMock, context, '/home/user', projectTargets, false, fsHost);
+    await deployToFunction(
+      firebaseMock,
+      context,
+      workspaceRoot,
+      STATIC_BUILD_TARGET,
+      SERVER_BUILD_TARGET,
+      { preview: false  },
+      undefined,
+      fsHost
+    );
 
     expect(spy).toHaveBeenCalledTimes(1);
 
     const packageArgs = spy.calls.argsFor(0);
 
     expect(packageArgs).toEqual([
-      'dist/dist/browser/index.html',
-      'dist/dist/browser/index.original.html'
+      join(workspaceRoot, 'dist', 'dist', 'browser', 'index.html'),
+      join(workspaceRoot, 'dist', 'dist', 'browser', 'index.original.html')
+    ]);
+  });
+
+  it('should rename the index.html file in the nested dist (new)', async () => {
+    const spy = spyOn(fsHost, 'renameSync');
+    await deployToFunction(
+      firebaseMock,
+      context,
+      workspaceRoot,
+      STATIC_BUILD_TARGET,
+      SERVER_BUILD_TARGET,
+      { preview: false, outputPath: join('dist', 'functions') },
+      undefined,
+      fsHost
+    );
+
+    expect(spy).toHaveBeenCalledTimes(1);
+
+    const packageArgs = spy.calls.argsFor(0);
+
+    expect(packageArgs).toEqual([
+      join(workspaceRoot, 'dist', 'functions', 'dist', 'browser', 'index.html'),
+      join(workspaceRoot, 'dist', 'functions', 'dist', 'browser', 'index.original.html')
     ]);
   });
 
   it('should invoke firebase.deploy', async () => {
     const spy = spyOn(firebaseMock, 'deploy');
-    await deployToFunction(firebaseMock, context, '/home/user', projectTargets, false, fsHost);
+    await deployToFunction(
+      firebaseMock,
+      context,
+      workspaceRoot,
+      STATIC_BUILD_TARGET,
+      SERVER_BUILD_TARGET,
+      { preview: false },
+      undefined,
+      fsHost
+    );
 
     expect(spy).toHaveBeenCalledTimes(1);
   });
