@@ -1,64 +1,65 @@
 import { BuilderContext, BuilderOutput, createBuilder } from '@angular-devkit/architect';
-import { NodeJsSyncHost } from '@angular-devkit/core/node';
-import deploy from './actions';
-import { experimental, json, normalize } from '@angular-devkit/core';
-import { BuildTarget, DeployBuilderSchema } from '../interfaces';
-import { getFirebaseProjectName } from '../utils';
-
-type DeployBuilderOptions = DeployBuilderSchema & json.JsonObject;
+import deploy, { DeployBuilderOptions } from './actions';
+import { BuildTarget } from '../interfaces';
+import { getFirebaseProjectNameFromFs } from '../utils';
+import { getFirebaseTools } from '../firebaseTools';
 
 // Call the createBuilder() function to create a builder. This mirrors
 // createJobHandler() but add typings specific to Architect Builders.
-export default createBuilder<any>(
+export default createBuilder(
   async (options: DeployBuilderOptions, context: BuilderContext): Promise<BuilderOutput> => {
-    // The project root is added to a BuilderContext.
-    const root = normalize(context.workspaceRoot);
-    const workspace = new experimental.workspace.Workspace(
-      root,
-      new NodeJsSyncHost()
-    );
-    await workspace
-      .loadWorkspaceFromHost(normalize('angular.json'))
-      .toPromise();
-
     if (!context.target) {
       throw new Error('Cannot deploy the application without a target');
     }
 
-    const projectTargets = workspace.getProjectTargets(context.target.project);
-
-    const firebaseProject = getFirebaseProjectName(
+    const [defaultFirebaseProject, defulatFirebaseHostingSite] = getFirebaseProjectNameFromFs(
       context.workspaceRoot,
       context.target.project
     );
 
+    const firebaseProject = options.firebaseProject || defaultFirebaseProject;
     if (!firebaseProject) {
-      throw new Error('Cannot find firebase project for your app in .firebaserc');
+      throw new Error('Cannot detirmine the Firebase Project from your angular.json or .firebaserc');
+    }
+    if (firebaseProject !== defaultFirebaseProject) {
+      throw new Error('The Firebase Project specified by your angular.json or .firebaserc is in conflict');
     }
 
-    const buildTarget = options.buildTarget || `${context.target.project}:build:production`;
+    const firebaseHostingSite = options.firebaseHostingSite || defulatFirebaseHostingSite;
+    if (!firebaseHostingSite) {
+      throw new Error(`Cannot detirmine the Firebase Hosting Site from your angular.json or .firebaserc`);
+    }
+    if (firebaseHostingSite !== defulatFirebaseHostingSite) {
+      throw new Error('The Firebase Hosting Site specified by your angular.json or .firebaserc is in conflict');
+    }
 
-    const targets: BuildTarget[] = [{
-      name: buildTarget
-    }];
+    const staticBuildTarget = { name: options.browserTarget || options.buildTarget || `${context.target.project}:build:production` };
+
+    let prerenderBuildTarget: BuildTarget | undefined;
+    if (options.prerender) {
+      prerenderBuildTarget = {
+        name: options.prerenderTarget || `${context.target.project}:prerender:production`
+      };
+    }
+
+    let serverBuildTarget: BuildTarget | undefined;
     if (options.ssr) {
-      targets.push({
-        name: options.universalBuildTarget || `${context.target.project}:server:production`,
-        options: {
-          bundleDependencies: 'all'
-        }
-      });
+      serverBuildTarget = {
+        name: options.serverTarget || options.universalBuildTarget || `${context.target.project}:server:production`
+      };
     }
 
     try {
+      process.env.FIREBASE_DEPLOY_AGENT = 'angularfire';
       await deploy(
-        require('firebase-tools'),
+        (await getFirebaseTools()),
         context,
-        projectTargets,
-        targets,
+        staticBuildTarget,
+        serverBuildTarget,
+        prerenderBuildTarget,
         firebaseProject,
-        !!options.ssr,
-        !!options.preview
+        options,
+        process.env.FIREBASE_TOKEN,
       );
     } catch (e) {
       console.error('Error when trying to deploy: ');
