@@ -1,67 +1,30 @@
-import { Inject, Injectable, InjectionToken, NgZone, Optional, PLATFORM_ID } from '@angular/core';
-import { createStorageRef } from './ref';
-import { Observable } from 'rxjs';
-import {
-  FIREBASE_APP_NAME,
-  FIREBASE_OPTIONS,
-  FirebaseAppConfig,
-  FirebaseOptions,
-  ɵAngularFireSchedulers,
-  ɵfirebaseAppFactory,
-  ɵkeepUnstableUntilFirstFactory
-} from '@angular/fire';
-import { UploadMetadata } from './interfaces';
-import { storage } from 'firebase/app';
-import 'firebase/storage';
-import firebase from 'firebase/app';
-import { registerStorage } from '@firebase/storage';
+import { FirebaseStorage } from 'firebase/storage';
+import { ɵgetAllInstancesOf } from '@angular/fire';
+import { from, timer } from 'rxjs';
+import { concatMap, distinct } from 'rxjs/operators';
 
-export const BUCKET = new InjectionToken<string>('angularfire2.storageBucket');
+// see notes in core/firebase.app.module.ts for why we're building the class like this
+// tslint:disable-next-line:no-empty-interface
+export interface Storage extends FirebaseStorage {}
 
-/**
- * AngularFireStorage Service
- *
- * This service is the main entry point for this feature module. It provides
- * an API for uploading and downloading binary files from Cloud Storage for
- * Firebase.
- */
-@Injectable({
-  providedIn: 'any'
-})
-export class AngularFireStorage {
-  public readonly storage: storage.Storage;
-
-  public readonly keepUnstableUntilFirst: <T>(obs: Observable<T>) => Observable<T>;
-  public readonly schedulers: ɵAngularFireSchedulers;
-
-  constructor(
-    @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
-    @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig: string | FirebaseAppConfig | null | undefined,
-    @Optional() @Inject(BUCKET) storageBucket: string | null,
-    // tslint:disable-next-line:ban-types
-    @Inject(PLATFORM_ID) platformId: Object,
-    zone: NgZone
-  ) {
-    this.schedulers = new ɵAngularFireSchedulers(zone);
-    this.keepUnstableUntilFirst = ɵkeepUnstableUntilFirstFactory(this.schedulers);
-
-    this.storage = zone.runOutsideAngular(() => {
-      const app = ɵfirebaseAppFactory(options, zone, nameOrConfig);
-      if (registerStorage) {
-        registerStorage(firebase as any);
-      }
-      return app.storage(storageBucket || undefined);
-    });
+export class Storage {
+  constructor(auth: FirebaseStorage) {
+    return auth;
   }
-
-  ref(path: string) {
-    return createStorageRef(this.storage.ref(path), this.schedulers, this.keepUnstableUntilFirst);
-  }
-
-  upload(path: string, data: any, metadata?: UploadMetadata) {
-    const storageRef = this.storage.ref(path);
-    const ref = createStorageRef(storageRef, this.schedulers, this.keepUnstableUntilFirst);
-    return ref.put(data, metadata);
-  }
-
 }
+
+export const STORAGE_PROVIDER_NAME = 'storage';
+
+// tslint:disable-next-line:no-empty-interface
+export interface StorageInstances extends Array<FirebaseStorage> {}
+
+export class StorageInstances {
+  constructor() {
+    return ɵgetAllInstancesOf<FirebaseStorage>(STORAGE_PROVIDER_NAME);
+  }
+}
+
+export const storageInstance$ = timer(0, 300).pipe(
+  concatMap(() => from(ɵgetAllInstancesOf<FirebaseStorage>(STORAGE_PROVIDER_NAME))),
+  distinct(),
+);
